Add files via upload
This is the primary implementation of one of the proposed distributed machine learning methods called "Group Alternating Direct Method of Multipliers (GADMM)". Here is the notebook file where the algorithm has been implemented considering 10 users. Anyone interested can extend the number of users. But, it takes more time when the iteration and the number of users increases.